pyproject.toml
[project]
name = "kappe"
authors = [{ name = "Marko Bausch" }]
description = "Kappe is an efficient data migration tool designed to seamlessly convert and split MCAP files."
readme = "README.md"
requires-python = ">=3.10"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "mcap>=0.3.0",
    "mcap-ros1-support>=0.6.0",
    "mcap-ros2-support>=0.3.0",
    "numpy",
    "pydantic>=2.4.2",
    "strictyaml",
    "tqdm",
    "pyyaml>=6.0.1",
    "scipy>=1.11.1",
    "jsonargparse[signatures]>=4.25.0",
]

dynamic = ["version"]

[project.urls]
repository = "https://github.com/sensmore/kappe"

[project.scripts]
kappe = "kappe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.version]
path = "src/kappe/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
exclude = ['msgs']
select = ['ALL']

ignore = [
    'ANN101',  # missing-type-self
    'ANN401',  # dynamically-typed-expression
    'C901',    # complex-structure
    'D',       # pydocstyle
    'EM101',   # raw-string-in-exception
    'EM102',   # f-string-in-exception
    'FIX002',  # TODO
    'PLR2004', # MagicValueComparisons
    'S101',    # AssertUsed
    'TD',      # TODO
    'TRY003',  # raise-vanilla-args

    # formater
    'COM812', # missing-trailing-comma
    'ISC001', # single-line-implicit-string-concatenation
]

src = ['src']

target-version = 'py310'

[tool.ruff.flake8-quotes]
inline-quotes = 'single'

[tool.ruff.format]
quote-style = "single"
